state: add Binding() to Volume and Filesystem
We add the Binding() method to Volume and Filesystem,
backed by a new "binding" document field. This field
identifies an entity whose destruction triggers the
destruction of the volume or filesystem.

This is the mechanism we will use to ensure that, by
default, volumes and filesystems will be destroyed
when their associated storage is destroyed. Volumes
that back filesystems are bound to the filesystem;
volumes and filesystems directly related to storage
instances will be bound to said storage; otherwise
volumes and filesystems may be bound to the environment
(destroy when the environment is (cleanly) destroyed),
to the machine they're attached to (destroy on detach),
or to nothing (don't destroy ever, unless explicitly
requested).

+++ state/filesystem.go
@@ -33,6 +33,27 @@ type Filesystem interface {
 	// Life returns the life of the filesystem.
 	Life() Life
 
+	// Binding is the tag of an entity that the filesystem's life span is
+	// bound to. This may be nil (i.e. the filesystem may persist beyond
+	// the span of environment), or one of EnvironmentTag, StorageTag, or
+	// MachineTag. The filesystem will be destroyed when the identified
+	// entity is destroyed.
+	//
+	// The following is a description of the effects of binding to
+	// different entities:
+	//   Machine:     If the filesystem is bound to a machine, then the
+	//                filesystem will be destroyed when it is detached from
+	//                the machine. It is not permitted for a filesystem to
+	//                be attached to multiple machines while it is bound to
+	//                a machine.
+	//   Storage:     If the filesystem is bound to a storage instance,
+	//                then the filesystem will be destroyed when the
+	//                storage insance is removed from state.
+	//   Environment: If the filesystem is bound to the environment, then
+	//                the filesystem must be destroyed prior to the
+	//                environment being destroyed.
+	Binding() names.Tag
+
 	// Storage returns the tag of the storage instance that this
 	// filesystem is assigned to, if any. If the filesystem is not
 	// assigned to a storage instance, an error satisfying
@@ -100,8 +121,10 @@ type filesystemDoc struct {
 	StorageId    string `bson:"storageid,omitempty"`
 	VolumeId     string `bson:"volumeid,omitempty"`
 	// TODO(axw) 2015-06-22 #1467379
-	// upgrade step to set this for 1.24 environments
+	// upgrade step to set "attachmentcount" and "binding"
+	// for 1.24 environments.
 	AttachmentCount int               `bson:"attachmentcount"`
+	Binding         string            `bson:"binding,omitempty"`
 	Info            *FilesystemInfo   `bson:"info,omitempty"`
 	Params          *FilesystemParams `bson:"params,omitempty"`
 }
@@ -123,6 +146,10 @@ type FilesystemParams struct {
 	// storage, if non-zero, is the tag of the storage instance
 	// that the filesystem is to be assigned to.
 	storage names.StorageTag
+
+	// binding, if non-nil, is the tag of the entity to which
+	// the filesystem's lifecycle will be bound.
+	binding names.Tag
 
 	Pool string `bson:"pool"`
 	Size uint64 `bson:"size"`
@@ -155,6 +182,26 @@ type FilesystemAttachmentParams struct {
 	ReadOnly bool   `bson:"read-only"`
 }
 
+// validate validates the contents of the filesystem document.
+func (f *filesystem) validate() error {
+	if f.doc.Binding != "" {
+		tag, err := names.ParseTag(f.doc.Binding)
+		if err != nil {
+			return errors.Annotate(err, "parsing binding")
+		}
+		switch tag.(type) {
+		case names.EnvironTag:
+			// TODO(axw) support binding to environment
+			return errors.NotSupportedf("binding to environment")
+		case names.MachineTag:
+		case names.StorageTag:
+		default:
+			return errors.Errorf("invalid binding: %v", f.doc.Binding)
+		}
+	}
+	return nil
+}
+
 // Tag is required to implement Entity.
 func (f *filesystem) Tag() names.Tag {
 	return f.FilesystemTag()
@@ -168,6 +215,16 @@ func (f *filesystem) FilesystemTag() names.FilesystemTag {
 // Life is required to implement Filesystem.
 func (f *filesystem) Life() Life {
 	return f.doc.Life
+}
+
+// Binding is required to implement Filesystem.
+func (f *filesystem) Binding() names.Tag {
+	if f.doc.Binding == "" {
+		return nil
+	}
+	// Tag is validated in filesystem.validate.
+	tag, _ := names.ParseTag(f.doc.Binding)
+	return tag
 }
 
 // Storage is required to implement Filesystem.
@@ -236,30 +293,38 @@ func (f *filesystemAttachment) Params() (FilesystemAttachmentParams, bool) {
 
 // Filesystem returns the Filesystem with the specified name.
 func (st *State) Filesystem(tag names.FilesystemTag) (Filesystem, error) {
-	f, err := st.filesystemInternal(tag)
+	f, err := st.filesystemByTag(tag)
 	return f, err
 }
 
-func (st *State) filesystemInternal(tag names.FilesystemTag) (*filesystem, error) {
+func (st *State) filesystemByTag(tag names.FilesystemTag) (*filesystem, error) {
 	query := bson.D{{"_id", tag.Id()}}
-	description := names.ReadableString(tag)
+	description := fmt.Sprintf("filesystem %q", tag.Id())
+	return st.filesystem(query, description)
+}
+
+func (st *State) storageInstanceFilesystem(tag names.StorageTag) (*filesystem, error) {
+	query := bson.D{{"storageid", tag.Id()}}
+	description := fmt.Sprintf("filesystem for storage instance %q", tag.Id())
 	return st.filesystem(query, description)
 }
 
 // StorageInstanceFilesystem returns the Filesystem assigned to the specified
 // storage instance.
 func (st *State) StorageInstanceFilesystem(tag names.StorageTag) (Filesystem, error) {
-	query := bson.D{{"storageid", tag.Id()}}
-	description := fmt.Sprintf("filesystem for storage instance %q", tag.Id())
-	f, err := st.filesystem(query, description)
+	f, err := st.storageInstanceFilesystem(tag)
 	return f, err
+}
+
+func (st *State) volumeFilesystem(tag names.VolumeTag) (*filesystem, error) {
+	query := bson.D{{"volumeid", tag.Id()}}
+	description := fmt.Sprintf("filesystem for volume %q", tag.Id())
+	return st.filesystem(query, description)
 }
 
 // VolumeFilesystem returns the Filesystem backed by the specified volume.
 func (st *State) VolumeFilesystem(tag names.VolumeTag) (Filesystem, error) {
-	query := bson.D{{"volumeid", tag.Id()}}
-	description := fmt.Sprintf("filesystem for volume %q", tag.Id())
-	f, err := st.filesystem(query, description)
+	f, err := st.volumeFilesystem(tag)
 	return f, err
 }
 
@@ -273,6 +338,9 @@ func (st *State) filesystem(query bson.D, description string) (*filesystem, erro
 		return nil, errors.NotFoundf(description)
 	} else if err != nil {
 		return nil, errors.Annotate(err, "cannot get filesystem")
+	}
+	if err := f.validate(); err != nil {
+		return nil, errors.Annotate(err, "validating filesystem")
 	}
 	return &f, nil
 }
@@ -434,7 +502,7 @@ func (st *State) RemoveFilesystemAttachment(machine names.MachineTag, filesystem
 		if attachment.Life() != Dying {
 			return nil, errors.New("filesystem attachment is not dying")
 		}
-		f, err := st.filesystemInternal(filesystem)
+		f, err := st.filesystemByTag(filesystem)
 		if err != nil {
 			return nil, errors.Trace(err)
 		}
@@ -459,6 +527,7 @@ func removeFilesystemAttachmentOps(m names.MachineTag, f *filesystem) []txn.Op {
 	decrefFilesystemOp := machineStorageDecrefOp(
 		filesystemsC, f.doc.FilesystemId,
 		f.doc.AttachmentCount, f.doc.Life,
+		m, f.doc.Binding,
 	)
 	return []txn.Op{{
 		C:      filesystemAttachmentsC,
@@ -472,7 +541,7 @@ func removeFilesystemAttachmentOps(m names.MachineTag, f *filesystem) []txn.Op {
 // be destroyed and removed from state at some point in the future.
 func (st *State) DestroyFilesystem(tag names.FilesystemTag) (err error) {
 	buildTxn := func(attempt int) ([]txn.Op, error) {
-		filesystem, err := st.filesystemInternal(tag)
+		filesystem, err := st.filesystemByTag(tag)
 		if errors.IsNotFound(err) {
 			return nil, jujutxn.ErrNoOperations
 		} else if err != nil {
@@ -534,19 +603,18 @@ func removeFilesystemOps(st *State, filesystem Filesystem) ([]txn.Op, error) {
 		Assert: txn.DocExists,
 		Remove: true,
 	}}
-	// TODO(axw) The below should only be done if the volume's
-	// lifecycle is bound to the filesystem, or to the machine
-	// that the filesystem is scoped to and the machine is Dying.
-	//
-	// If the filesystem is backed by a volume, the volume can and
-	// should be destroyed once the filesystem is removed.
+	// If the filesystem is backed by a volume, the volume should
+	// be destroyed once the filesystem is removed if it is bound
+	// to the filesystem.
 	volumeTag, err := filesystem.Volume()
 	if err == nil {
-		volume, err := st.volume(volumeTag)
+		volume, err := st.volumeByTag(volumeTag)
 		if err != nil {
 			return nil, errors.Trace(err)
 		}
-		ops = append(ops, destroyVolumeOps(st, volume)...)
+		if volume.Binding() == filesystem.Tag() {
+			ops = append(ops, destroyVolumeOps(st, volume)...)
+		}
 	} else if err != ErrNoBackingVolume {
 		return nil, errors.Trace(err)
 	}
@@ -592,6 +660,9 @@ func newFilesystemId(st *State, machineId string) (string, error) {
 // directly, a volume will be created and Juju will manage a filesystem
 // on it.
 func (st *State) addFilesystemOps(params FilesystemParams, machineId string) ([]txn.Op, names.FilesystemTag, names.VolumeTag, error) {
+	if params.binding == nil {
+		params.binding = names.NewMachineTag(machineId)
+	}
 	params, err := st.filesystemParamsWithDefaults(params)
 	if err != nil {
 		return nil, names.FilesystemTag{}, names.VolumeTag{}, errors.Trace(err)
@@ -600,6 +671,12 @@ func (st *State) addFilesystemOps(params FilesystemParams, machineId string) ([]
 	if err != nil {
 		return nil, names.FilesystemTag{}, names.VolumeTag{}, errors.Annotate(err, "validating filesystem params")
 	}
+
+	filesystemId, err := newFilesystemId(st, machineId)
+	if err != nil {
+		return nil, names.FilesystemTag{}, names.VolumeTag{}, errors.Annotate(err, "cannot generate filesystem name")
+	}
+	filesystemTag := names.NewFilesystemTag(filesystemId)
 
 	// Check if the filesystem needs a volume.
 	var volumeId string
@@ -613,6 +690,7 @@ func (st *State) addFilesystemOps(params FilesystemParams, machineId string) ([]
 		var volumeOp txn.Op
 		volumeParams := VolumeParams{
 			params.storage,
+			filesystemTag, // volume is bound to filesystem
 			params.Pool,
 			params.Size,
 		}
@@ -624,25 +702,22 @@ func (st *State) addFilesystemOps(params FilesystemParams, machineId string) ([]
 		ops = append(ops, volumeOp)
 	}
 
-	id, err := newFilesystemId(st, machineId)
-	if err != nil {
-		return nil, names.FilesystemTag{}, names.VolumeTag{}, errors.Annotate(err, "cannot generate filesystem name")
-	}
 	filesystemOp := txn.Op{
 		C:      filesystemsC,
-		Id:     id,
+		Id:     filesystemId,
 		Assert: txn.DocMissing,
 		Insert: &filesystemDoc{
-			FilesystemId: id,
+			FilesystemId: filesystemId,
 			VolumeId:     volumeId,
 			StorageId:    params.storage.Id(),
+			Binding:      params.binding.String(),
 			Params:       &params,
 			// Every filesystem is created with one attachment.
 			AttachmentCount: 1,
 		},
 	}
 	ops = append(ops, filesystemOp)
-	return ops, names.NewFilesystemTag(id), volumeTag, nil
+	return ops, filesystemTag, volumeTag, nil
 }
 
 func (st *State) filesystemParamsWithDefaults(params FilesystemParams) (FilesystemParams, error) {
